+++ rust-src/concordium_base/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 - `cbor::cbor_encode` is now infallible and returns `Vec<u8>` instead of `CborSerializationResult<Vec<u8>>`
+- `&[T]` no longer implements `CborSerialize` in order to avoid ambiguity with `CborSerialize` implementation for `[u8]`. 
+  To serialize a slice as a CBOR array, wrap it in `ArraySlice`.
+  `[u8]` still implements `CborSerialize` and serializes as a CBOR bytestring.
 - Removed the module `upward`. This will be added to Rust SDK crate instead. 
 - Changes to `protocol_level_tokens` module:
   - Removed the usage of `Upward` in `TokenOperations` type. To CBOR decode and allow unknown variants,
@@ -18,8 +21,7 @@
     `TokenModuleRejectReason::encode_reject_reason` and `TokenModuleRejectReason::decode_reject_reason` allows CBOR encoding from and CBOR decoding
     to `TokenModuleRejectReason`. The new methods `TokenModuleRejectReasonType::to_type_discriminator` and
     `TokenModuleRejectReasonType::try_from_type_discriminator` allows converting between `TokenModuleRejectReasonType`
-    and `TokenModuleCborTypeDiscriminator`. 
-
+    and `TokenModuleCborTypeDiscriminator`.
 
 - The flag `serde_deprecated` now guards `serde::Serialize` and `serde::Deserialize` implemetations on the following types. The implementations will eventually be removed.
   - `protocol_level_tokens::token_metadata_url::MetadataUrl`